@@ -8,10 +8,10 @@
         }
     ],
     "require": {
-        "php": "~7.0"
+        "php": "~7.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~6.0"
+        "phpunit/phpunit": "~7.0"
     },
     "autoload": {
         "psr-4": {
